Split TrackingPointMethods to enable sharing
